requirements.txt
streamlit
pandas
numpy
scikit-learn
shap
matplotlib
fpdf
Pillow